TECH-1121 : add workflow management helper + clean up + linter on PR
