+++ fixtures/graphql/jcr/mutation/setPropertyValues.graphql
@@ -1,0 +1,9 @@
+mutation setProperty($pathOrId: String!, $property: String!, $value: [String]!, $language: String!){
+    jcr{
+        mutateNode(pathOrId: $pathOrId){
+            mutateProperty(name: $property){
+                setValues(values: $value, language: $language)
+            }
+        }
+    }
+}
