+++ src/custom.d.ts
@@ -1,3 +1,4 @@
 interface String {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     replaceAll(input: string, output : string): any;
 }
