+++ src/support/fixture.ts
@@ -1,9 +1,9 @@
 import Chainable = Cypress.Chainable;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const fixture = function (originalCommand: ((...args: any[]) => any), fixture: string, ...args: any[]): Chainable<any> {
+export const fixture = function (originalCommand: ((...args: any[]) => any), fixtureParam: string, ...args: any[]): Chainable<any> {
     return cy.wrap({}, {log: false}).then(() => {
-        return originalCommand(fixture, ...args).then(f => {
+        return originalCommand(fixtureParam, ...args).then(f => {
             return f;
         }).catch(() => {
             return null;
@@ -16,7 +16,7 @@ export const fixture = function (originalCommand: ((...args: any[]) => any), fix
             }
 
             try {
-                cy.readFile('./node_modules/@jahia/cypress/fixtures/' + fixture, encoding, {log: false, timeout: 200});
+                cy.readFile('./node_modules/@jahia/cypress/fixtures/' + fixtureParam, encoding, {log: false, timeout: 200});
             } catch (e) {
                 console.log(e);
             }
