+++ src/utils/UsersHelper.ts
@@ -1,0 +1,30 @@
+
+export const grantRoles = (pathOrId: string, roleNames: Array<string>, principalName: string, principalType: string): Cypress.Chainable => {
+    cy.log('Grant role(s) ' + roleNames + ' with principal type ' + principalType + ' to ' + principalName + ' on node ' + pathOrId);
+    return cy.apollo({
+        variables: {
+            pathOrId: pathOrId,
+            roleNames: roleNames,
+            principalName: principalName,
+            principalType: principalType
+        },
+        mutationFile: 'graphql/jcr/mutation/grantRoles.graphql'
+    });
+};
+
+export const createUser = (userName: string, password: string, properties: { name: string, value: string }[] = []): void => {
+    const userProperties = properties.map(property => {
+        return 'properties.setProperty("' + property.name + '", "' + property.value + '")';
+    });
+    cy.executeGroovy('groovy/admin/createUser.groovy', {
+        USER_NAME: userName,
+        PASSWORD: password ? password : 'password',
+        USER_PROPERTIES: userProperties ? userProperties.join('\n') : ''
+    });
+};
+
+export const deleteUser = (userName: string): void => {
+    cy.executeGroovy('groovy/admin/deleteUser.groovy', {
+        USER_NAME: userName
+    });
+};
